+++ 12/5.c
@@ -4,13 +4,13 @@
 
 int main()
 {
-    float r, a, b;
-    scanf("%f", &r);
+    double r, a, b;
+    scanf("%lf", &r);
 
     a = 2 * PI * r;
     b = PI * r * r;
-    printf("%.3f\n", a);
-    printf("%.3f", b);
+    printf("%.3lf\n", a);
+    printf("%.3lf", b);
 
     return 0;
 }
